Updated simulation to be more in tune with what a real-simulation would be: Iterated progress based on the creatures attributes. Still much more to be done to generate a richer simulation.
Updated the natural selection process to care only about the carrying capacity instead of the current population size. This is because, in a large population it tends to be that the top creatures survive and the bottom creatures die, however in a small population all creatures tend to survive. This leads me to believe it has nothing to do with the population size, but rather the carrying capacity of the environment. Thus I removed the need to gradient over the population size.